change  check directory. when kill listener process, close check_jobs.sh

+++ check_jobs.sh
@@ -2,18 +2,18 @@
 # Check jobs acceptance timeout.
 # If jobs cannot be confirmed, listening will be terminated.
 
-
-_RUNNER_WORKDIR=${RUNNER_WORKDIR:-/_work}
 _JOBS_ACCEPTANCE_TIMEOUT=$1
+_CHECK_DIR=/actions-runner/_diag/pages
 
-while [ 1 ]; do
-  if [ ! -e ${_RUNNER_WORKDIR}/_temp/_github_workflow ]; then
-    _DURATION=`expr $(date +%s) - $(date +%s -r /tmp)`
-    [[ ${_DURATION} -ge ${_JOBS_ACCEPTANCE_TIMEOUT} ]] && break
-  fi
+sleep 1
+while [[ -n `pgrep Runner.Listener` ]]; do
   sleep 10
-done
+  _DURATION=`expr $(date +%s) - $(date +%s -r /tmp)`
+  [[ -e ${_CHECK_DIR} ]] && [[ -n "$(ls -A ${_CHECK_DIR})" ]] && continue
+  [[ ${_DURATION} -lt ${_JOBS_ACCEPTANCE_TIMEOUT} ]] && continue
 
-echo $(date '+%Y-%m-%d %H:%M:%SZ:') Stop listening due to timeout.
-pkill --signal=SIGINT -f ./bin/Runner.Listener
+  echo $(date '+%Y-%m-%d %H:%M:%SZ:') Stop listening due to timeout.
+  pkill --signal=SIGINT -f ./bin/Runner.Listener
+  break
+done
 exit 0
